\name{DocTypeExtensions}
\alias{DocTypeExtensions}
\docType{data}
\title{Collection of extension and MIME-type pairs for Google Documents}
\description{
  This is a named vector giving the pairs of
  extensions and corresponding file/MIME types.
  These are used when  uploading a document.
}
%\usage{data(DocTypeExtensions)}
\format{
  The format is:
 chr "DocTypeExtensions"
}
\details{
}
\source{

}
\references{

}
\examples{

}
\keyword{datasets}
